+++ opt-env-conf-test/test_resources/docs/same-help/nix-options.nix
@@ -1,0 +1,18 @@
+{ lib }:
+{
+  int = lib.mkOption {
+    default = null;
+    description = "int or string";
+    type = lib.types.nullOr lib.types.int;
+  };
+  other = lib.mkOption {
+    default = null;
+    description = "int or string";
+    type = lib.types.nullOr lib.types.bool;
+  };
+  string = lib.mkOption {
+    default = null;
+    description = "int or string";
+    type = lib.types.nullOr lib.types.str;
+  };
+}